Update jquery.tokeninput.js
minChars option now works with a value of 0. If minChars is set to 0, a search will be performed when the input gains focus.

placeHolderText has been added as an option. When set, the value in the placeHolderText variable will appear as placeholder text in the input.

extraParams has been added as an option. This can either be an object (set of key value pairs) or a function which returns such an object. The keys are the query param names and the values are their corresponding values.