Update jquery.tokeninput.js
Fixed a bug involving the resizing of the input box when the last token is deleted. The placeholder used to stay hidden. Now the placeholder appears again when the input is empty.

I also changed the condition on when the results are displayed. Now they are only displayed in an input if the input still has the focus.