Add option to control whether first result is automatically selected
Currently, the first result in the drop down is automatically selected. This makes free free tagging difficult because it defaults to that value even if the user intended to enter just a substring of that result.

This commit adds an autoSelectFirstResult option which allows to disable that behavior. If it is set to false (the default value is set to true to not break backwards compatibility), the first result won't be automatically selected. Instead, the user has to select it himself.