Added maximum length attribute option
When sending a key term to a database by specifying a data type (i.e.
table column data type) in stored procedure, if a user enters more than
a certain number of characters, it causes an error. So, I added the
option to limit the number of characters to the search text box.