Improving keyboard management
Improving keyboard management

Issues :
* On Safari iPad, the <input> is not updated until keydown() has ended,
despite setTimeout(), which causes do_search() to perform over the
previous value of the field.
* Keypress event returns "l" (lowercase "L", ASCII 108) as keyCode,
which is then confused for KEY.NUMPAD_ENTER (keyCode 108).

Fixes :
* Updating the KEY array to the latest jQuery UI standard :
https://github.com/jquery/jquery-ui/blob/e9643f6bfc922773e29e3084e64de0240b547f32/ui/core.js
* Replacing event.keyCode by jQuery event.which :
https://api.jquery.com/event.which/
* Dissociating keydown/keyup effects :
http://www.quirksmode.org/dom/events/keys.html
** Keydown event is used to detect control keys (when the key is
pressed, meaning that the effect of the key can be overloaded before it
happens)
** Keyup event is used to act upon input changes (when the key is
released, meaning that the effect of the key has already happened)
** Keypress event is not used anymore
** Now that do_search() performs only on keyup, the setTimeout() ("just
long enough to let this function finish.") seems useless.
* Removed stopPropagation(), preventDefault() when redundant with
"return false"